+++ src/components/Wheel.tsx
@@ -151,7 +151,7 @@ export const Wheel = React.memo(
       if (renderer === 'skia') {
         if (SkiaComponent == null) {
           throw new Error(
-            '[@vap/react-native-wheel] renderer="skia" requires @shopify/react-native-skia to be installed.'
+            '[@vapng/react-native-wheel] renderer="skia" requires @shopify/react-native-skia to be installed.'
           );
         }
         return (